requirements.txt
# GastroAssist AI Dependencies

# Backend Framework
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2

# API Middleware
starlette>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.0
psycopg2-binary>=2.9.7  # For PostgreSQL connection

# LLM Integration
openai>=1.2.4
langchain>=0.0.335
langchain-community>=0.0.10

# Vector Storage
pinecone-client>=2.2.4
hnswlib>=0.7.0

# Search Integrations
tavily-python>=0.2.2
duckduckgo-search>=3.9.3

# Data Processing
numpy>=1.25.2
pandas>=2.1.0
unstructured>=0.10.8
pypdf>=3.15.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.24.1

# Utilities
tenacity>=8.2.3
tiktoken>=0.5.1
tqdm>=4.66.1
